Wrap toast in hook and renaming
