+++ src/components/Toast/ToastProvider.tsx
@@ -1,13 +1,13 @@
 import React, { FC, useState } from 'react';
 import { LayoutAnimation } from 'react-native';
 import { ToastContext } from 'src/components/Toast/toastContext';
-import { ToastProps } from 'src/components/Toast/toastTypes';
+import { ToastProps } from 'src/components/Toast/types';
 
 export const ToastProvider: FC = ({ children }) => {
   const [toast, setToast] = useState<ToastProps>();
   const DEFAULT_DURATION = 5000;
 
-  const displayToast = (options: ToastProps) => {
+  const addToast = (options: ToastProps) => {
     LayoutAnimation.configureNext(LayoutAnimation.Presets.easeInEaseOut);
     setToast(options);
 
@@ -22,8 +22,8 @@ export const ToastProvider: FC = ({ children }) => {
   };
 
   return (
-    <ToastContext.Provider value={{ toast, displayToast, closeToast }}>
-      { children}
+    <ToastContext.Provider value={{ toast, addToast, closeToast }}>
+      {children}
     </ToastContext.Provider>
   );
 };
