+++ src/components/Toast/index.tsx
@@ -2,4 +2,4 @@ export * from './Toast';
 export * from './toastContext';
 export * from './ToastProvider';
 export * from './ToastRenderer';
-export * from './toastTypes';
+export * from './types';
