--- src/components/Toast/toastTypes.ts
@@ -1,10 +1,0 @@
-export type ToastProps = {
-  message: string;
-  duration?: number;
-};
-
-export type ToastContextProps = {
-  toast?: ToastProps;
-  displayToast: (options: ToastProps) => void;
-  closeToast: () => void;
-};
