+++ src/components/Toast/types.ts
@@ -1,0 +1,14 @@
+export type ToastProps = {
+  message: string;
+  duration?: number;
+};
+
+export interface ToastItem extends ToastProps {
+  id: number;
+}
+
+export type ToastOptions = {
+  toast?: ToastProps;
+  addToast: (options: ToastProps) => void;
+  closeToast: () => void;
+};
